refactor: extract shared extractCategory function
Consolidate three duplicate extractCategory implementations into a
single shared pure function in gen2-migration/categories.ts.

Changes:
- Create categories.ts with extractCategory() supporting 12 categories
- Update drift-formatter.ts to use shared function
- Update _validations.ts to use shared function
- Remove extractCategory from amplify-config-service.ts (unused)

This improves testability and eliminates divergence risk between
duplicate implementations.
---
Prompt: Extract extractCategory to shared pure function in
gen2-migration/categories.ts, eliminating duplicate implementations
in drift-formatter.ts, amplify-config-service.ts, and _validations.ts
